Seed engine

This helper function is used inside of plane_seed to evaluate characteristics of observed data to use for downstream plausibility analysis.

Usage

seed_engine(input, location, cut_date = NULL)

Arguments

input

Input signal data used for seeding; must be an observed signal object

location

Character vector with location code

cut_date

Maximum date (inclusive) for which seeding should be performed; default is NULL and the entire input will be used for seeding

Value

A list of length 1 with multiple elements corresponding to seed characteristics and metadata for the given location.