Score PLANES components

This function wraps PLANES scoring for specified components across all locations in single step.

Usage

plane_score(input, seed, components = "all", args = NULL)

Arguments

input

Input signal data to be scored; object must be one of forecast or observed

seed

Prepared seed

components

Character vector specifying components. Must be either "all" or any combination of "cover", "diff", "taper", "trend", and "repeats". Default is 'all' and will use all available components for the given signal

args

Named list of arguments for component functions. List elements must be named to match the given component and arguments passed as a nested list (e.g., args = list(trend = list(sig_lvl = 0.05))). Default is NULL and defaults for all components will be used

Value

A list with scoring results for all locations.

Examples

## read in example observed data and prep observed signal
hosp <- read.csv(system.file("extdata/observed/hdgov_hosp_weekly.csv", package = "rplanes"))
hosp$date <- as.Date(hosp$date, format = "%Y-%m-%d")
prepped_observed <- to_signal(hosp, outcome = "flu.admits", type = "observed", resolution = "weeks")

## read in example forecast and prep forecast signal
fp <- system.file("extdata/forecast/2022-10-31-SigSci-TSENS.csv", package = "rplanes")
prepped_forecast <- read_forecast(fp) %>%
  to_signal(., outcome = "flu.admits", type = "forecast", horizon = 4)

## prepare seed with cut date
prepped_seed <- plane_seed(prepped_observed, cut_date = "2022-10-29")

## run plane scoring with all components
plane_score(input = prepped_forecast, seed = prepped_seed)
#> $scores_summary
#> $scores_summary$`02`
#> $scores_summary$`02`$location
#> [1] "02"
#> 
#> $scores_summary$`02`$n_flags
#> [1] 1
#> 
#> $scores_summary$`02`$n_components
#> [1] 5
#> 
#> $scores_summary$`02`$score
#> [1] 0.2
#> 
#> $scores_summary$`02`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`04`
#> $scores_summary$`04`$location
#> [1] "04"
#> 
#> $scores_summary$`04`$n_flags
#> [1] 0
#> 
#> $scores_summary$`04`$n_components
#> [1] 5
#> 
#> $scores_summary$`04`$score
#> [1] 0
#> 
#> $scores_summary$`04`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`05`
#> $scores_summary$`05`$location
#> [1] "05"
#> 
#> $scores_summary$`05`$n_flags
#> [1] 0
#> 
#> $scores_summary$`05`$n_components
#> [1] 5
#> 
#> $scores_summary$`05`$score
#> [1] 0
#> 
#> $scores_summary$`05`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`06`
#> $scores_summary$`06`$location
#> [1] "06"
#> 
#> $scores_summary$`06`$n_flags
#> [1] 0
#> 
#> $scores_summary$`06`$n_components
#> [1] 5
#> 
#> $scores_summary$`06`$score
#> [1] 0
#> 
#> $scores_summary$`06`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`08`
#> $scores_summary$`08`$location
#> [1] "08"
#> 
#> $scores_summary$`08`$n_flags
#> [1] 0
#> 
#> $scores_summary$`08`$n_components
#> [1] 5
#> 
#> $scores_summary$`08`$score
#> [1] 0
#> 
#> $scores_summary$`08`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`09`
#> $scores_summary$`09`$location
#> [1] "09"
#> 
#> $scores_summary$`09`$n_flags
#> [1] 3
#> 
#> $scores_summary$`09`$n_components
#> [1] 5
#> 
#> $scores_summary$`09`$score
#> [1] 0.6
#> 
#> $scores_summary$`09`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`10`
#> $scores_summary$`10`$location
#> [1] "10"
#> 
#> $scores_summary$`10`$n_flags
#> [1] 0
#> 
#> $scores_summary$`10`$n_components
#> [1] 5
#> 
#> $scores_summary$`10`$score
#> [1] 0
#> 
#> $scores_summary$`10`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`12`
#> $scores_summary$`12`$location
#> [1] "12"
#> 
#> $scores_summary$`12`$n_flags
#> [1] 0
#> 
#> $scores_summary$`12`$n_components
#> [1] 5
#> 
#> $scores_summary$`12`$score
#> [1] 0
#> 
#> $scores_summary$`12`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`13`
#> $scores_summary$`13`$location
#> [1] "13"
#> 
#> $scores_summary$`13`$n_flags
#> [1] 0
#> 
#> $scores_summary$`13`$n_components
#> [1] 5
#> 
#> $scores_summary$`13`$score
#> [1] 0
#> 
#> $scores_summary$`13`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`15`
#> $scores_summary$`15`$location
#> [1] "15"
#> 
#> $scores_summary$`15`$n_flags
#> [1] 1
#> 
#> $scores_summary$`15`$n_components
#> [1] 5
#> 
#> $scores_summary$`15`$score
#> [1] 0.2
#> 
#> $scores_summary$`15`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`16`
#> $scores_summary$`16`$location
#> [1] "16"
#> 
#> $scores_summary$`16`$n_flags
#> [1] 0
#> 
#> $scores_summary$`16`$n_components
#> [1] 5
#> 
#> $scores_summary$`16`$score
#> [1] 0
#> 
#> $scores_summary$`16`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`17`
#> $scores_summary$`17`$location
#> [1] "17"
#> 
#> $scores_summary$`17`$n_flags
#> [1] 0
#> 
#> $scores_summary$`17`$n_components
#> [1] 5
#> 
#> $scores_summary$`17`$score
#> [1] 0
#> 
#> $scores_summary$`17`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`18`
#> $scores_summary$`18`$location
#> [1] "18"
#> 
#> $scores_summary$`18`$n_flags
#> [1] 0
#> 
#> $scores_summary$`18`$n_components
#> [1] 5
#> 
#> $scores_summary$`18`$score
#> [1] 0
#> 
#> $scores_summary$`18`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`19`
#> $scores_summary$`19`$location
#> [1] "19"
#> 
#> $scores_summary$`19`$n_flags
#> [1] 1
#> 
#> $scores_summary$`19`$n_components
#> [1] 5
#> 
#> $scores_summary$`19`$score
#> [1] 0.2
#> 
#> $scores_summary$`19`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`20`
#> $scores_summary$`20`$location
#> [1] "20"
#> 
#> $scores_summary$`20`$n_flags
#> [1] 0
#> 
#> $scores_summary$`20`$n_components
#> [1] 5
#> 
#> $scores_summary$`20`$score
#> [1] 0
#> 
#> $scores_summary$`20`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`21`
#> $scores_summary$`21`$location
#> [1] "21"
#> 
#> $scores_summary$`21`$n_flags
#> [1] 0
#> 
#> $scores_summary$`21`$n_components
#> [1] 5
#> 
#> $scores_summary$`21`$score
#> [1] 0
#> 
#> $scores_summary$`21`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`22`
#> $scores_summary$`22`$location
#> [1] "22"
#> 
#> $scores_summary$`22`$n_flags
#> [1] 1
#> 
#> $scores_summary$`22`$n_components
#> [1] 5
#> 
#> $scores_summary$`22`$score
#> [1] 0.2
#> 
#> $scores_summary$`22`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`23`
#> $scores_summary$`23`$location
#> [1] "23"
#> 
#> $scores_summary$`23`$n_flags
#> [1] 0
#> 
#> $scores_summary$`23`$n_components
#> [1] 5
#> 
#> $scores_summary$`23`$score
#> [1] 0
#> 
#> $scores_summary$`23`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`24`
#> $scores_summary$`24`$location
#> [1] "24"
#> 
#> $scores_summary$`24`$n_flags
#> [1] 1
#> 
#> $scores_summary$`24`$n_components
#> [1] 5
#> 
#> $scores_summary$`24`$score
#> [1] 0.2
#> 
#> $scores_summary$`24`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`25`
#> $scores_summary$`25`$location
#> [1] "25"
#> 
#> $scores_summary$`25`$n_flags
#> [1] 0
#> 
#> $scores_summary$`25`$n_components
#> [1] 5
#> 
#> $scores_summary$`25`$score
#> [1] 0
#> 
#> $scores_summary$`25`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`26`
#> $scores_summary$`26`$location
#> [1] "26"
#> 
#> $scores_summary$`26`$n_flags
#> [1] 0
#> 
#> $scores_summary$`26`$n_components
#> [1] 5
#> 
#> $scores_summary$`26`$score
#> [1] 0
#> 
#> $scores_summary$`26`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`27`
#> $scores_summary$`27`$location
#> [1] "27"
#> 
#> $scores_summary$`27`$n_flags
#> [1] 0
#> 
#> $scores_summary$`27`$n_components
#> [1] 5
#> 
#> $scores_summary$`27`$score
#> [1] 0
#> 
#> $scores_summary$`27`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`28`
#> $scores_summary$`28`$location
#> [1] "28"
#> 
#> $scores_summary$`28`$n_flags
#> [1] 0
#> 
#> $scores_summary$`28`$n_components
#> [1] 5
#> 
#> $scores_summary$`28`$score
#> [1] 0
#> 
#> $scores_summary$`28`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`29`
#> $scores_summary$`29`$location
#> [1] "29"
#> 
#> $scores_summary$`29`$n_flags
#> [1] 0
#> 
#> $scores_summary$`29`$n_components
#> [1] 5
#> 
#> $scores_summary$`29`$score
#> [1] 0
#> 
#> $scores_summary$`29`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`30`
#> $scores_summary$`30`$location
#> [1] "30"
#> 
#> $scores_summary$`30`$n_flags
#> [1] 0
#> 
#> $scores_summary$`30`$n_components
#> [1] 5
#> 
#> $scores_summary$`30`$score
#> [1] 0
#> 
#> $scores_summary$`30`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`31`
#> $scores_summary$`31`$location
#> [1] "31"
#> 
#> $scores_summary$`31`$n_flags
#> [1] 0
#> 
#> $scores_summary$`31`$n_components
#> [1] 5
#> 
#> $scores_summary$`31`$score
#> [1] 0
#> 
#> $scores_summary$`31`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`32`
#> $scores_summary$`32`$location
#> [1] "32"
#> 
#> $scores_summary$`32`$n_flags
#> [1] 1
#> 
#> $scores_summary$`32`$n_components
#> [1] 5
#> 
#> $scores_summary$`32`$score
#> [1] 0.2
#> 
#> $scores_summary$`32`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`33`
#> $scores_summary$`33`$location
#> [1] "33"
#> 
#> $scores_summary$`33`$n_flags
#> [1] 0
#> 
#> $scores_summary$`33`$n_components
#> [1] 5
#> 
#> $scores_summary$`33`$score
#> [1] 0
#> 
#> $scores_summary$`33`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`34`
#> $scores_summary$`34`$location
#> [1] "34"
#> 
#> $scores_summary$`34`$n_flags
#> [1] 0
#> 
#> $scores_summary$`34`$n_components
#> [1] 5
#> 
#> $scores_summary$`34`$score
#> [1] 0
#> 
#> $scores_summary$`34`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`35`
#> $scores_summary$`35`$location
#> [1] "35"
#> 
#> $scores_summary$`35`$n_flags
#> [1] 0
#> 
#> $scores_summary$`35`$n_components
#> [1] 5
#> 
#> $scores_summary$`35`$score
#> [1] 0
#> 
#> $scores_summary$`35`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`36`
#> $scores_summary$`36`$location
#> [1] "36"
#> 
#> $scores_summary$`36`$n_flags
#> [1] 0
#> 
#> $scores_summary$`36`$n_components
#> [1] 5
#> 
#> $scores_summary$`36`$score
#> [1] 0
#> 
#> $scores_summary$`36`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`37`
#> $scores_summary$`37`$location
#> [1] "37"
#> 
#> $scores_summary$`37`$n_flags
#> [1] 1
#> 
#> $scores_summary$`37`$n_components
#> [1] 5
#> 
#> $scores_summary$`37`$score
#> [1] 0.2
#> 
#> $scores_summary$`37`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`38`
#> $scores_summary$`38`$location
#> [1] "38"
#> 
#> $scores_summary$`38`$n_flags
#> [1] 0
#> 
#> $scores_summary$`38`$n_components
#> [1] 5
#> 
#> $scores_summary$`38`$score
#> [1] 0
#> 
#> $scores_summary$`38`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`39`
#> $scores_summary$`39`$location
#> [1] "39"
#> 
#> $scores_summary$`39`$n_flags
#> [1] 0
#> 
#> $scores_summary$`39`$n_components
#> [1] 5
#> 
#> $scores_summary$`39`$score
#> [1] 0
#> 
#> $scores_summary$`39`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`40`
#> $scores_summary$`40`$location
#> [1] "40"
#> 
#> $scores_summary$`40`$n_flags
#> [1] 0
#> 
#> $scores_summary$`40`$n_components
#> [1] 5
#> 
#> $scores_summary$`40`$score
#> [1] 0
#> 
#> $scores_summary$`40`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`41`
#> $scores_summary$`41`$location
#> [1] "41"
#> 
#> $scores_summary$`41`$n_flags
#> [1] 0
#> 
#> $scores_summary$`41`$n_components
#> [1] 5
#> 
#> $scores_summary$`41`$score
#> [1] 0
#> 
#> $scores_summary$`41`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`42`
#> $scores_summary$`42`$location
#> [1] "42"
#> 
#> $scores_summary$`42`$n_flags
#> [1] 0
#> 
#> $scores_summary$`42`$n_components
#> [1] 5
#> 
#> $scores_summary$`42`$score
#> [1] 0
#> 
#> $scores_summary$`42`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`44`
#> $scores_summary$`44`$location
#> [1] "44"
#> 
#> $scores_summary$`44`$n_flags
#> [1] 0
#> 
#> $scores_summary$`44`$n_components
#> [1] 5
#> 
#> $scores_summary$`44`$score
#> [1] 0
#> 
#> $scores_summary$`44`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`45`
#> $scores_summary$`45`$location
#> [1] "45"
#> 
#> $scores_summary$`45`$n_flags
#> [1] 2
#> 
#> $scores_summary$`45`$n_components
#> [1] 5
#> 
#> $scores_summary$`45`$score
#> [1] 0.4
#> 
#> $scores_summary$`45`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`46`
#> $scores_summary$`46`$location
#> [1] "46"
#> 
#> $scores_summary$`46`$n_flags
#> [1] 0
#> 
#> $scores_summary$`46`$n_components
#> [1] 5
#> 
#> $scores_summary$`46`$score
#> [1] 0
#> 
#> $scores_summary$`46`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`47`
#> $scores_summary$`47`$location
#> [1] "47"
#> 
#> $scores_summary$`47`$n_flags
#> [1] 1
#> 
#> $scores_summary$`47`$n_components
#> [1] 5
#> 
#> $scores_summary$`47`$score
#> [1] 0.2
#> 
#> $scores_summary$`47`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`48`
#> $scores_summary$`48`$location
#> [1] "48"
#> 
#> $scores_summary$`48`$n_flags
#> [1] 1
#> 
#> $scores_summary$`48`$n_components
#> [1] 5
#> 
#> $scores_summary$`48`$score
#> [1] 0.2
#> 
#> $scores_summary$`48`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`49`
#> $scores_summary$`49`$location
#> [1] "49"
#> 
#> $scores_summary$`49`$n_flags
#> [1] 0
#> 
#> $scores_summary$`49`$n_components
#> [1] 5
#> 
#> $scores_summary$`49`$score
#> [1] 0
#> 
#> $scores_summary$`49`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`50`
#> $scores_summary$`50`$location
#> [1] "50"
#> 
#> $scores_summary$`50`$n_flags
#> [1] 0
#> 
#> $scores_summary$`50`$n_components
#> [1] 5
#> 
#> $scores_summary$`50`$score
#> [1] 0
#> 
#> $scores_summary$`50`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`51`
#> $scores_summary$`51`$location
#> [1] "51"
#> 
#> $scores_summary$`51`$n_flags
#> [1] 0
#> 
#> $scores_summary$`51`$n_components
#> [1] 5
#> 
#> $scores_summary$`51`$score
#> [1] 0
#> 
#> $scores_summary$`51`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`53`
#> $scores_summary$`53`$location
#> [1] "53"
#> 
#> $scores_summary$`53`$n_flags
#> [1] 0
#> 
#> $scores_summary$`53`$n_components
#> [1] 5
#> 
#> $scores_summary$`53`$score
#> [1] 0
#> 
#> $scores_summary$`53`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`54`
#> $scores_summary$`54`$location
#> [1] "54"
#> 
#> $scores_summary$`54`$n_flags
#> [1] 2
#> 
#> $scores_summary$`54`$n_components
#> [1] 5
#> 
#> $scores_summary$`54`$score
#> [1] 0.4
#> 
#> $scores_summary$`54`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`55`
#> $scores_summary$`55`$location
#> [1] "55"
#> 
#> $scores_summary$`55`$n_flags
#> [1] 0
#> 
#> $scores_summary$`55`$n_components
#> [1] 5
#> 
#> $scores_summary$`55`$score
#> [1] 0
#> 
#> $scores_summary$`55`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`56`
#> $scores_summary$`56`$location
#> [1] "56"
#> 
#> $scores_summary$`56`$n_flags
#> [1] 0
#> 
#> $scores_summary$`56`$n_components
#> [1] 5
#> 
#> $scores_summary$`56`$score
#> [1] 0
#> 
#> $scores_summary$`56`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$US
#> $scores_summary$US$location
#> [1] "US"
#> 
#> $scores_summary$US$n_flags
#> [1] 1
#> 
#> $scores_summary$US$n_components
#> [1] 5
#> 
#> $scores_summary$US$score
#> [1] 0.2
#> 
#> $scores_summary$US$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> 
#> $scores_raw
#> # A tibble: 250 × 3
#>    component location indicator
#>    <chr>     <chr>    <lgl>    
#>  1 cover     02       TRUE     
#>  2 diff      02       FALSE    
#>  3 repeats   02       FALSE    
#>  4 taper     02       FALSE    
#>  5 trend     02       FALSE    
#>  6 cover     04       FALSE    
#>  7 diff      04       FALSE    
#>  8 repeats   04       FALSE    
#>  9 taper     04       FALSE    
#> 10 trend     04       FALSE    
#> # ℹ 240 more rows
#> 

## run plane scoring with select components
plane_score(input = prepped_forecast, seed = prepped_seed, components = c("cover","taper"))
#> $scores_summary
#> $scores_summary$`02`
#> $scores_summary$`02`$location
#> [1] "02"
#> 
#> $scores_summary$`02`$n_flags
#> [1] 1
#> 
#> $scores_summary$`02`$n_components
#> [1] 2
#> 
#> $scores_summary$`02`$score
#> [1] 0.5
#> 
#> $scores_summary$`02`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`04`
#> $scores_summary$`04`$location
#> [1] "04"
#> 
#> $scores_summary$`04`$n_flags
#> [1] 0
#> 
#> $scores_summary$`04`$n_components
#> [1] 2
#> 
#> $scores_summary$`04`$score
#> [1] 0
#> 
#> $scores_summary$`04`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`05`
#> $scores_summary$`05`$location
#> [1] "05"
#> 
#> $scores_summary$`05`$n_flags
#> [1] 0
#> 
#> $scores_summary$`05`$n_components
#> [1] 2
#> 
#> $scores_summary$`05`$score
#> [1] 0
#> 
#> $scores_summary$`05`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`06`
#> $scores_summary$`06`$location
#> [1] "06"
#> 
#> $scores_summary$`06`$n_flags
#> [1] 0
#> 
#> $scores_summary$`06`$n_components
#> [1] 2
#> 
#> $scores_summary$`06`$score
#> [1] 0
#> 
#> $scores_summary$`06`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`08`
#> $scores_summary$`08`$location
#> [1] "08"
#> 
#> $scores_summary$`08`$n_flags
#> [1] 0
#> 
#> $scores_summary$`08`$n_components
#> [1] 2
#> 
#> $scores_summary$`08`$score
#> [1] 0
#> 
#> $scores_summary$`08`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`09`
#> $scores_summary$`09`$location
#> [1] "09"
#> 
#> $scores_summary$`09`$n_flags
#> [1] 1
#> 
#> $scores_summary$`09`$n_components
#> [1] 2
#> 
#> $scores_summary$`09`$score
#> [1] 0.5
#> 
#> $scores_summary$`09`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`10`
#> $scores_summary$`10`$location
#> [1] "10"
#> 
#> $scores_summary$`10`$n_flags
#> [1] 0
#> 
#> $scores_summary$`10`$n_components
#> [1] 2
#> 
#> $scores_summary$`10`$score
#> [1] 0
#> 
#> $scores_summary$`10`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`12`
#> $scores_summary$`12`$location
#> [1] "12"
#> 
#> $scores_summary$`12`$n_flags
#> [1] 0
#> 
#> $scores_summary$`12`$n_components
#> [1] 2
#> 
#> $scores_summary$`12`$score
#> [1] 0
#> 
#> $scores_summary$`12`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`13`
#> $scores_summary$`13`$location
#> [1] "13"
#> 
#> $scores_summary$`13`$n_flags
#> [1] 0
#> 
#> $scores_summary$`13`$n_components
#> [1] 2
#> 
#> $scores_summary$`13`$score
#> [1] 0
#> 
#> $scores_summary$`13`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`15`
#> $scores_summary$`15`$location
#> [1] "15"
#> 
#> $scores_summary$`15`$n_flags
#> [1] 1
#> 
#> $scores_summary$`15`$n_components
#> [1] 2
#> 
#> $scores_summary$`15`$score
#> [1] 0.5
#> 
#> $scores_summary$`15`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`16`
#> $scores_summary$`16`$location
#> [1] "16"
#> 
#> $scores_summary$`16`$n_flags
#> [1] 0
#> 
#> $scores_summary$`16`$n_components
#> [1] 2
#> 
#> $scores_summary$`16`$score
#> [1] 0
#> 
#> $scores_summary$`16`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`17`
#> $scores_summary$`17`$location
#> [1] "17"
#> 
#> $scores_summary$`17`$n_flags
#> [1] 0
#> 
#> $scores_summary$`17`$n_components
#> [1] 2
#> 
#> $scores_summary$`17`$score
#> [1] 0
#> 
#> $scores_summary$`17`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`18`
#> $scores_summary$`18`$location
#> [1] "18"
#> 
#> $scores_summary$`18`$n_flags
#> [1] 0
#> 
#> $scores_summary$`18`$n_components
#> [1] 2
#> 
#> $scores_summary$`18`$score
#> [1] 0
#> 
#> $scores_summary$`18`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`19`
#> $scores_summary$`19`$location
#> [1] "19"
#> 
#> $scores_summary$`19`$n_flags
#> [1] 0
#> 
#> $scores_summary$`19`$n_components
#> [1] 2
#> 
#> $scores_summary$`19`$score
#> [1] 0
#> 
#> $scores_summary$`19`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`20`
#> $scores_summary$`20`$location
#> [1] "20"
#> 
#> $scores_summary$`20`$n_flags
#> [1] 0
#> 
#> $scores_summary$`20`$n_components
#> [1] 2
#> 
#> $scores_summary$`20`$score
#> [1] 0
#> 
#> $scores_summary$`20`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`21`
#> $scores_summary$`21`$location
#> [1] "21"
#> 
#> $scores_summary$`21`$n_flags
#> [1] 0
#> 
#> $scores_summary$`21`$n_components
#> [1] 2
#> 
#> $scores_summary$`21`$score
#> [1] 0
#> 
#> $scores_summary$`21`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`22`
#> $scores_summary$`22`$location
#> [1] "22"
#> 
#> $scores_summary$`22`$n_flags
#> [1] 1
#> 
#> $scores_summary$`22`$n_components
#> [1] 2
#> 
#> $scores_summary$`22`$score
#> [1] 0.5
#> 
#> $scores_summary$`22`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`23`
#> $scores_summary$`23`$location
#> [1] "23"
#> 
#> $scores_summary$`23`$n_flags
#> [1] 0
#> 
#> $scores_summary$`23`$n_components
#> [1] 2
#> 
#> $scores_summary$`23`$score
#> [1] 0
#> 
#> $scores_summary$`23`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`24`
#> $scores_summary$`24`$location
#> [1] "24"
#> 
#> $scores_summary$`24`$n_flags
#> [1] 0
#> 
#> $scores_summary$`24`$n_components
#> [1] 2
#> 
#> $scores_summary$`24`$score
#> [1] 0
#> 
#> $scores_summary$`24`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`25`
#> $scores_summary$`25`$location
#> [1] "25"
#> 
#> $scores_summary$`25`$n_flags
#> [1] 0
#> 
#> $scores_summary$`25`$n_components
#> [1] 2
#> 
#> $scores_summary$`25`$score
#> [1] 0
#> 
#> $scores_summary$`25`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`26`
#> $scores_summary$`26`$location
#> [1] "26"
#> 
#> $scores_summary$`26`$n_flags
#> [1] 0
#> 
#> $scores_summary$`26`$n_components
#> [1] 2
#> 
#> $scores_summary$`26`$score
#> [1] 0
#> 
#> $scores_summary$`26`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`27`
#> $scores_summary$`27`$location
#> [1] "27"
#> 
#> $scores_summary$`27`$n_flags
#> [1] 0
#> 
#> $scores_summary$`27`$n_components
#> [1] 2
#> 
#> $scores_summary$`27`$score
#> [1] 0
#> 
#> $scores_summary$`27`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`28`
#> $scores_summary$`28`$location
#> [1] "28"
#> 
#> $scores_summary$`28`$n_flags
#> [1] 0
#> 
#> $scores_summary$`28`$n_components
#> [1] 2
#> 
#> $scores_summary$`28`$score
#> [1] 0
#> 
#> $scores_summary$`28`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`29`
#> $scores_summary$`29`$location
#> [1] "29"
#> 
#> $scores_summary$`29`$n_flags
#> [1] 0
#> 
#> $scores_summary$`29`$n_components
#> [1] 2
#> 
#> $scores_summary$`29`$score
#> [1] 0
#> 
#> $scores_summary$`29`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`30`
#> $scores_summary$`30`$location
#> [1] "30"
#> 
#> $scores_summary$`30`$n_flags
#> [1] 0
#> 
#> $scores_summary$`30`$n_components
#> [1] 2
#> 
#> $scores_summary$`30`$score
#> [1] 0
#> 
#> $scores_summary$`30`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`31`
#> $scores_summary$`31`$location
#> [1] "31"
#> 
#> $scores_summary$`31`$n_flags
#> [1] 0
#> 
#> $scores_summary$`31`$n_components
#> [1] 2
#> 
#> $scores_summary$`31`$score
#> [1] 0
#> 
#> $scores_summary$`31`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`32`
#> $scores_summary$`32`$location
#> [1] "32"
#> 
#> $scores_summary$`32`$n_flags
#> [1] 0
#> 
#> $scores_summary$`32`$n_components
#> [1] 2
#> 
#> $scores_summary$`32`$score
#> [1] 0
#> 
#> $scores_summary$`32`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`33`
#> $scores_summary$`33`$location
#> [1] "33"
#> 
#> $scores_summary$`33`$n_flags
#> [1] 0
#> 
#> $scores_summary$`33`$n_components
#> [1] 2
#> 
#> $scores_summary$`33`$score
#> [1] 0
#> 
#> $scores_summary$`33`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`34`
#> $scores_summary$`34`$location
#> [1] "34"
#> 
#> $scores_summary$`34`$n_flags
#> [1] 0
#> 
#> $scores_summary$`34`$n_components
#> [1] 2
#> 
#> $scores_summary$`34`$score
#> [1] 0
#> 
#> $scores_summary$`34`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`35`
#> $scores_summary$`35`$location
#> [1] "35"
#> 
#> $scores_summary$`35`$n_flags
#> [1] 0
#> 
#> $scores_summary$`35`$n_components
#> [1] 2
#> 
#> $scores_summary$`35`$score
#> [1] 0
#> 
#> $scores_summary$`35`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`36`
#> $scores_summary$`36`$location
#> [1] "36"
#> 
#> $scores_summary$`36`$n_flags
#> [1] 0
#> 
#> $scores_summary$`36`$n_components
#> [1] 2
#> 
#> $scores_summary$`36`$score
#> [1] 0
#> 
#> $scores_summary$`36`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`37`
#> $scores_summary$`37`$location
#> [1] "37"
#> 
#> $scores_summary$`37`$n_flags
#> [1] 1
#> 
#> $scores_summary$`37`$n_components
#> [1] 2
#> 
#> $scores_summary$`37`$score
#> [1] 0.5
#> 
#> $scores_summary$`37`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`38`
#> $scores_summary$`38`$location
#> [1] "38"
#> 
#> $scores_summary$`38`$n_flags
#> [1] 0
#> 
#> $scores_summary$`38`$n_components
#> [1] 2
#> 
#> $scores_summary$`38`$score
#> [1] 0
#> 
#> $scores_summary$`38`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`39`
#> $scores_summary$`39`$location
#> [1] "39"
#> 
#> $scores_summary$`39`$n_flags
#> [1] 0
#> 
#> $scores_summary$`39`$n_components
#> [1] 2
#> 
#> $scores_summary$`39`$score
#> [1] 0
#> 
#> $scores_summary$`39`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`40`
#> $scores_summary$`40`$location
#> [1] "40"
#> 
#> $scores_summary$`40`$n_flags
#> [1] 0
#> 
#> $scores_summary$`40`$n_components
#> [1] 2
#> 
#> $scores_summary$`40`$score
#> [1] 0
#> 
#> $scores_summary$`40`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`41`
#> $scores_summary$`41`$location
#> [1] "41"
#> 
#> $scores_summary$`41`$n_flags
#> [1] 0
#> 
#> $scores_summary$`41`$n_components
#> [1] 2
#> 
#> $scores_summary$`41`$score
#> [1] 0
#> 
#> $scores_summary$`41`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`42`
#> $scores_summary$`42`$location
#> [1] "42"
#> 
#> $scores_summary$`42`$n_flags
#> [1] 0
#> 
#> $scores_summary$`42`$n_components
#> [1] 2
#> 
#> $scores_summary$`42`$score
#> [1] 0
#> 
#> $scores_summary$`42`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`44`
#> $scores_summary$`44`$location
#> [1] "44"
#> 
#> $scores_summary$`44`$n_flags
#> [1] 0
#> 
#> $scores_summary$`44`$n_components
#> [1] 2
#> 
#> $scores_summary$`44`$score
#> [1] 0
#> 
#> $scores_summary$`44`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`45`
#> $scores_summary$`45`$location
#> [1] "45"
#> 
#> $scores_summary$`45`$n_flags
#> [1] 1
#> 
#> $scores_summary$`45`$n_components
#> [1] 2
#> 
#> $scores_summary$`45`$score
#> [1] 0.5
#> 
#> $scores_summary$`45`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`46`
#> $scores_summary$`46`$location
#> [1] "46"
#> 
#> $scores_summary$`46`$n_flags
#> [1] 0
#> 
#> $scores_summary$`46`$n_components
#> [1] 2
#> 
#> $scores_summary$`46`$score
#> [1] 0
#> 
#> $scores_summary$`46`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`47`
#> $scores_summary$`47`$location
#> [1] "47"
#> 
#> $scores_summary$`47`$n_flags
#> [1] 0
#> 
#> $scores_summary$`47`$n_components
#> [1] 2
#> 
#> $scores_summary$`47`$score
#> [1] 0
#> 
#> $scores_summary$`47`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`48`
#> $scores_summary$`48`$location
#> [1] "48"
#> 
#> $scores_summary$`48`$n_flags
#> [1] 0
#> 
#> $scores_summary$`48`$n_components
#> [1] 2
#> 
#> $scores_summary$`48`$score
#> [1] 0
#> 
#> $scores_summary$`48`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`49`
#> $scores_summary$`49`$location
#> [1] "49"
#> 
#> $scores_summary$`49`$n_flags
#> [1] 0
#> 
#> $scores_summary$`49`$n_components
#> [1] 2
#> 
#> $scores_summary$`49`$score
#> [1] 0
#> 
#> $scores_summary$`49`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`50`
#> $scores_summary$`50`$location
#> [1] "50"
#> 
#> $scores_summary$`50`$n_flags
#> [1] 0
#> 
#> $scores_summary$`50`$n_components
#> [1] 2
#> 
#> $scores_summary$`50`$score
#> [1] 0
#> 
#> $scores_summary$`50`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`51`
#> $scores_summary$`51`$location
#> [1] "51"
#> 
#> $scores_summary$`51`$n_flags
#> [1] 0
#> 
#> $scores_summary$`51`$n_components
#> [1] 2
#> 
#> $scores_summary$`51`$score
#> [1] 0
#> 
#> $scores_summary$`51`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`53`
#> $scores_summary$`53`$location
#> [1] "53"
#> 
#> $scores_summary$`53`$n_flags
#> [1] 0
#> 
#> $scores_summary$`53`$n_components
#> [1] 2
#> 
#> $scores_summary$`53`$score
#> [1] 0
#> 
#> $scores_summary$`53`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`54`
#> $scores_summary$`54`$location
#> [1] "54"
#> 
#> $scores_summary$`54`$n_flags
#> [1] 1
#> 
#> $scores_summary$`54`$n_components
#> [1] 2
#> 
#> $scores_summary$`54`$score
#> [1] 0.5
#> 
#> $scores_summary$`54`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`55`
#> $scores_summary$`55`$location
#> [1] "55"
#> 
#> $scores_summary$`55`$n_flags
#> [1] 0
#> 
#> $scores_summary$`55`$n_components
#> [1] 2
#> 
#> $scores_summary$`55`$score
#> [1] 0
#> 
#> $scores_summary$`55`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$`56`
#> $scores_summary$`56`$location
#> [1] "56"
#> 
#> $scores_summary$`56`$n_flags
#> [1] 0
#> 
#> $scores_summary$`56`$n_components
#> [1] 2
#> 
#> $scores_summary$`56`$score
#> [1] 0
#> 
#> $scores_summary$`56`$components
#> [1] "cover;taper"
#> 
#> 
#> $scores_summary$US
#> $scores_summary$US$location
#> [1] "US"
#> 
#> $scores_summary$US$n_flags
#> [1] 1
#> 
#> $scores_summary$US$n_components
#> [1] 2
#> 
#> $scores_summary$US$score
#> [1] 0.5
#> 
#> $scores_summary$US$components
#> [1] "cover;taper"
#> 
#> 
#> 
#> $scores_raw
#> # A tibble: 100 × 3
#>    component location indicator
#>    <chr>     <chr>    <lgl>    
#>  1 cover     02       TRUE     
#>  2 taper     02       FALSE    
#>  3 cover     04       FALSE    
#>  4 taper     04       FALSE    
#>  5 cover     05       FALSE    
#>  6 taper     05       FALSE    
#>  7 cover     06       FALSE    
#>  8 taper     06       FALSE    
#>  9 cover     08       FALSE    
#> 10 taper     08       FALSE    
#> # ℹ 90 more rows
#> 

## run plane scoring with all components and additional args
comp_args <- list(trend = list(sig_lvl = 0.05), repeats = list(prepend = 4, tolerance = 8))
plane_score(input = prepped_forecast, seed = prepped_seed, args = comp_args)
#> $scores_summary
#> $scores_summary$`02`
#> $scores_summary$`02`$location
#> [1] "02"
#> 
#> $scores_summary$`02`$n_flags
#> [1] 1
#> 
#> $scores_summary$`02`$n_components
#> [1] 5
#> 
#> $scores_summary$`02`$score
#> [1] 0.2
#> 
#> $scores_summary$`02`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`04`
#> $scores_summary$`04`$location
#> [1] "04"
#> 
#> $scores_summary$`04`$n_flags
#> [1] 0
#> 
#> $scores_summary$`04`$n_components
#> [1] 5
#> 
#> $scores_summary$`04`$score
#> [1] 0
#> 
#> $scores_summary$`04`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`05`
#> $scores_summary$`05`$location
#> [1] "05"
#> 
#> $scores_summary$`05`$n_flags
#> [1] 0
#> 
#> $scores_summary$`05`$n_components
#> [1] 5
#> 
#> $scores_summary$`05`$score
#> [1] 0
#> 
#> $scores_summary$`05`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`06`
#> $scores_summary$`06`$location
#> [1] "06"
#> 
#> $scores_summary$`06`$n_flags
#> [1] 0
#> 
#> $scores_summary$`06`$n_components
#> [1] 5
#> 
#> $scores_summary$`06`$score
#> [1] 0
#> 
#> $scores_summary$`06`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`08`
#> $scores_summary$`08`$location
#> [1] "08"
#> 
#> $scores_summary$`08`$n_flags
#> [1] 0
#> 
#> $scores_summary$`08`$n_components
#> [1] 5
#> 
#> $scores_summary$`08`$score
#> [1] 0
#> 
#> $scores_summary$`08`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`09`
#> $scores_summary$`09`$location
#> [1] "09"
#> 
#> $scores_summary$`09`$n_flags
#> [1] 2
#> 
#> $scores_summary$`09`$n_components
#> [1] 5
#> 
#> $scores_summary$`09`$score
#> [1] 0.4
#> 
#> $scores_summary$`09`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`10`
#> $scores_summary$`10`$location
#> [1] "10"
#> 
#> $scores_summary$`10`$n_flags
#> [1] 0
#> 
#> $scores_summary$`10`$n_components
#> [1] 5
#> 
#> $scores_summary$`10`$score
#> [1] 0
#> 
#> $scores_summary$`10`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`12`
#> $scores_summary$`12`$location
#> [1] "12"
#> 
#> $scores_summary$`12`$n_flags
#> [1] 0
#> 
#> $scores_summary$`12`$n_components
#> [1] 5
#> 
#> $scores_summary$`12`$score
#> [1] 0
#> 
#> $scores_summary$`12`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`13`
#> $scores_summary$`13`$location
#> [1] "13"
#> 
#> $scores_summary$`13`$n_flags
#> [1] 0
#> 
#> $scores_summary$`13`$n_components
#> [1] 5
#> 
#> $scores_summary$`13`$score
#> [1] 0
#> 
#> $scores_summary$`13`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`15`
#> $scores_summary$`15`$location
#> [1] "15"
#> 
#> $scores_summary$`15`$n_flags
#> [1] 1
#> 
#> $scores_summary$`15`$n_components
#> [1] 5
#> 
#> $scores_summary$`15`$score
#> [1] 0.2
#> 
#> $scores_summary$`15`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`16`
#> $scores_summary$`16`$location
#> [1] "16"
#> 
#> $scores_summary$`16`$n_flags
#> [1] 0
#> 
#> $scores_summary$`16`$n_components
#> [1] 5
#> 
#> $scores_summary$`16`$score
#> [1] 0
#> 
#> $scores_summary$`16`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`17`
#> $scores_summary$`17`$location
#> [1] "17"
#> 
#> $scores_summary$`17`$n_flags
#> [1] 0
#> 
#> $scores_summary$`17`$n_components
#> [1] 5
#> 
#> $scores_summary$`17`$score
#> [1] 0
#> 
#> $scores_summary$`17`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`18`
#> $scores_summary$`18`$location
#> [1] "18"
#> 
#> $scores_summary$`18`$n_flags
#> [1] 0
#> 
#> $scores_summary$`18`$n_components
#> [1] 5
#> 
#> $scores_summary$`18`$score
#> [1] 0
#> 
#> $scores_summary$`18`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`19`
#> $scores_summary$`19`$location
#> [1] "19"
#> 
#> $scores_summary$`19`$n_flags
#> [1] 0
#> 
#> $scores_summary$`19`$n_components
#> [1] 5
#> 
#> $scores_summary$`19`$score
#> [1] 0
#> 
#> $scores_summary$`19`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`20`
#> $scores_summary$`20`$location
#> [1] "20"
#> 
#> $scores_summary$`20`$n_flags
#> [1] 0
#> 
#> $scores_summary$`20`$n_components
#> [1] 5
#> 
#> $scores_summary$`20`$score
#> [1] 0
#> 
#> $scores_summary$`20`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`21`
#> $scores_summary$`21`$location
#> [1] "21"
#> 
#> $scores_summary$`21`$n_flags
#> [1] 0
#> 
#> $scores_summary$`21`$n_components
#> [1] 5
#> 
#> $scores_summary$`21`$score
#> [1] 0
#> 
#> $scores_summary$`21`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`22`
#> $scores_summary$`22`$location
#> [1] "22"
#> 
#> $scores_summary$`22`$n_flags
#> [1] 1
#> 
#> $scores_summary$`22`$n_components
#> [1] 5
#> 
#> $scores_summary$`22`$score
#> [1] 0.2
#> 
#> $scores_summary$`22`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`23`
#> $scores_summary$`23`$location
#> [1] "23"
#> 
#> $scores_summary$`23`$n_flags
#> [1] 0
#> 
#> $scores_summary$`23`$n_components
#> [1] 5
#> 
#> $scores_summary$`23`$score
#> [1] 0
#> 
#> $scores_summary$`23`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`24`
#> $scores_summary$`24`$location
#> [1] "24"
#> 
#> $scores_summary$`24`$n_flags
#> [1] 0
#> 
#> $scores_summary$`24`$n_components
#> [1] 5
#> 
#> $scores_summary$`24`$score
#> [1] 0
#> 
#> $scores_summary$`24`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`25`
#> $scores_summary$`25`$location
#> [1] "25"
#> 
#> $scores_summary$`25`$n_flags
#> [1] 0
#> 
#> $scores_summary$`25`$n_components
#> [1] 5
#> 
#> $scores_summary$`25`$score
#> [1] 0
#> 
#> $scores_summary$`25`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`26`
#> $scores_summary$`26`$location
#> [1] "26"
#> 
#> $scores_summary$`26`$n_flags
#> [1] 0
#> 
#> $scores_summary$`26`$n_components
#> [1] 5
#> 
#> $scores_summary$`26`$score
#> [1] 0
#> 
#> $scores_summary$`26`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`27`
#> $scores_summary$`27`$location
#> [1] "27"
#> 
#> $scores_summary$`27`$n_flags
#> [1] 0
#> 
#> $scores_summary$`27`$n_components
#> [1] 5
#> 
#> $scores_summary$`27`$score
#> [1] 0
#> 
#> $scores_summary$`27`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`28`
#> $scores_summary$`28`$location
#> [1] "28"
#> 
#> $scores_summary$`28`$n_flags
#> [1] 0
#> 
#> $scores_summary$`28`$n_components
#> [1] 5
#> 
#> $scores_summary$`28`$score
#> [1] 0
#> 
#> $scores_summary$`28`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`29`
#> $scores_summary$`29`$location
#> [1] "29"
#> 
#> $scores_summary$`29`$n_flags
#> [1] 0
#> 
#> $scores_summary$`29`$n_components
#> [1] 5
#> 
#> $scores_summary$`29`$score
#> [1] 0
#> 
#> $scores_summary$`29`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`30`
#> $scores_summary$`30`$location
#> [1] "30"
#> 
#> $scores_summary$`30`$n_flags
#> [1] 0
#> 
#> $scores_summary$`30`$n_components
#> [1] 5
#> 
#> $scores_summary$`30`$score
#> [1] 0
#> 
#> $scores_summary$`30`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`31`
#> $scores_summary$`31`$location
#> [1] "31"
#> 
#> $scores_summary$`31`$n_flags
#> [1] 0
#> 
#> $scores_summary$`31`$n_components
#> [1] 5
#> 
#> $scores_summary$`31`$score
#> [1] 0
#> 
#> $scores_summary$`31`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`32`
#> $scores_summary$`32`$location
#> [1] "32"
#> 
#> $scores_summary$`32`$n_flags
#> [1] 0
#> 
#> $scores_summary$`32`$n_components
#> [1] 5
#> 
#> $scores_summary$`32`$score
#> [1] 0
#> 
#> $scores_summary$`32`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`33`
#> $scores_summary$`33`$location
#> [1] "33"
#> 
#> $scores_summary$`33`$n_flags
#> [1] 0
#> 
#> $scores_summary$`33`$n_components
#> [1] 5
#> 
#> $scores_summary$`33`$score
#> [1] 0
#> 
#> $scores_summary$`33`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`34`
#> $scores_summary$`34`$location
#> [1] "34"
#> 
#> $scores_summary$`34`$n_flags
#> [1] 0
#> 
#> $scores_summary$`34`$n_components
#> [1] 5
#> 
#> $scores_summary$`34`$score
#> [1] 0
#> 
#> $scores_summary$`34`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`35`
#> $scores_summary$`35`$location
#> [1] "35"
#> 
#> $scores_summary$`35`$n_flags
#> [1] 0
#> 
#> $scores_summary$`35`$n_components
#> [1] 5
#> 
#> $scores_summary$`35`$score
#> [1] 0
#> 
#> $scores_summary$`35`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`36`
#> $scores_summary$`36`$location
#> [1] "36"
#> 
#> $scores_summary$`36`$n_flags
#> [1] 0
#> 
#> $scores_summary$`36`$n_components
#> [1] 5
#> 
#> $scores_summary$`36`$score
#> [1] 0
#> 
#> $scores_summary$`36`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`37`
#> $scores_summary$`37`$location
#> [1] "37"
#> 
#> $scores_summary$`37`$n_flags
#> [1] 1
#> 
#> $scores_summary$`37`$n_components
#> [1] 5
#> 
#> $scores_summary$`37`$score
#> [1] 0.2
#> 
#> $scores_summary$`37`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`38`
#> $scores_summary$`38`$location
#> [1] "38"
#> 
#> $scores_summary$`38`$n_flags
#> [1] 0
#> 
#> $scores_summary$`38`$n_components
#> [1] 5
#> 
#> $scores_summary$`38`$score
#> [1] 0
#> 
#> $scores_summary$`38`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`39`
#> $scores_summary$`39`$location
#> [1] "39"
#> 
#> $scores_summary$`39`$n_flags
#> [1] 0
#> 
#> $scores_summary$`39`$n_components
#> [1] 5
#> 
#> $scores_summary$`39`$score
#> [1] 0
#> 
#> $scores_summary$`39`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`40`
#> $scores_summary$`40`$location
#> [1] "40"
#> 
#> $scores_summary$`40`$n_flags
#> [1] 0
#> 
#> $scores_summary$`40`$n_components
#> [1] 5
#> 
#> $scores_summary$`40`$score
#> [1] 0
#> 
#> $scores_summary$`40`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`41`
#> $scores_summary$`41`$location
#> [1] "41"
#> 
#> $scores_summary$`41`$n_flags
#> [1] 0
#> 
#> $scores_summary$`41`$n_components
#> [1] 5
#> 
#> $scores_summary$`41`$score
#> [1] 0
#> 
#> $scores_summary$`41`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`42`
#> $scores_summary$`42`$location
#> [1] "42"
#> 
#> $scores_summary$`42`$n_flags
#> [1] 0
#> 
#> $scores_summary$`42`$n_components
#> [1] 5
#> 
#> $scores_summary$`42`$score
#> [1] 0
#> 
#> $scores_summary$`42`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`44`
#> $scores_summary$`44`$location
#> [1] "44"
#> 
#> $scores_summary$`44`$n_flags
#> [1] 0
#> 
#> $scores_summary$`44`$n_components
#> [1] 5
#> 
#> $scores_summary$`44`$score
#> [1] 0
#> 
#> $scores_summary$`44`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`45`
#> $scores_summary$`45`$location
#> [1] "45"
#> 
#> $scores_summary$`45`$n_flags
#> [1] 2
#> 
#> $scores_summary$`45`$n_components
#> [1] 5
#> 
#> $scores_summary$`45`$score
#> [1] 0.4
#> 
#> $scores_summary$`45`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`46`
#> $scores_summary$`46`$location
#> [1] "46"
#> 
#> $scores_summary$`46`$n_flags
#> [1] 0
#> 
#> $scores_summary$`46`$n_components
#> [1] 5
#> 
#> $scores_summary$`46`$score
#> [1] 0
#> 
#> $scores_summary$`46`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`47`
#> $scores_summary$`47`$location
#> [1] "47"
#> 
#> $scores_summary$`47`$n_flags
#> [1] 0
#> 
#> $scores_summary$`47`$n_components
#> [1] 5
#> 
#> $scores_summary$`47`$score
#> [1] 0
#> 
#> $scores_summary$`47`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`48`
#> $scores_summary$`48`$location
#> [1] "48"
#> 
#> $scores_summary$`48`$n_flags
#> [1] 0
#> 
#> $scores_summary$`48`$n_components
#> [1] 5
#> 
#> $scores_summary$`48`$score
#> [1] 0
#> 
#> $scores_summary$`48`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`49`
#> $scores_summary$`49`$location
#> [1] "49"
#> 
#> $scores_summary$`49`$n_flags
#> [1] 0
#> 
#> $scores_summary$`49`$n_components
#> [1] 5
#> 
#> $scores_summary$`49`$score
#> [1] 0
#> 
#> $scores_summary$`49`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`50`
#> $scores_summary$`50`$location
#> [1] "50"
#> 
#> $scores_summary$`50`$n_flags
#> [1] 0
#> 
#> $scores_summary$`50`$n_components
#> [1] 5
#> 
#> $scores_summary$`50`$score
#> [1] 0
#> 
#> $scores_summary$`50`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`51`
#> $scores_summary$`51`$location
#> [1] "51"
#> 
#> $scores_summary$`51`$n_flags
#> [1] 0
#> 
#> $scores_summary$`51`$n_components
#> [1] 5
#> 
#> $scores_summary$`51`$score
#> [1] 0
#> 
#> $scores_summary$`51`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`53`
#> $scores_summary$`53`$location
#> [1] "53"
#> 
#> $scores_summary$`53`$n_flags
#> [1] 0
#> 
#> $scores_summary$`53`$n_components
#> [1] 5
#> 
#> $scores_summary$`53`$score
#> [1] 0
#> 
#> $scores_summary$`53`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`54`
#> $scores_summary$`54`$location
#> [1] "54"
#> 
#> $scores_summary$`54`$n_flags
#> [1] 1
#> 
#> $scores_summary$`54`$n_components
#> [1] 5
#> 
#> $scores_summary$`54`$score
#> [1] 0.2
#> 
#> $scores_summary$`54`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`55`
#> $scores_summary$`55`$location
#> [1] "55"
#> 
#> $scores_summary$`55`$n_flags
#> [1] 0
#> 
#> $scores_summary$`55`$n_components
#> [1] 5
#> 
#> $scores_summary$`55`$score
#> [1] 0
#> 
#> $scores_summary$`55`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$`56`
#> $scores_summary$`56`$location
#> [1] "56"
#> 
#> $scores_summary$`56`$n_flags
#> [1] 0
#> 
#> $scores_summary$`56`$n_components
#> [1] 5
#> 
#> $scores_summary$`56`$score
#> [1] 0
#> 
#> $scores_summary$`56`$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> $scores_summary$US
#> $scores_summary$US$location
#> [1] "US"
#> 
#> $scores_summary$US$n_flags
#> [1] 1
#> 
#> $scores_summary$US$n_components
#> [1] 5
#> 
#> $scores_summary$US$score
#> [1] 0.2
#> 
#> $scores_summary$US$components
#> [1] "cover;diff;repeats;taper;trend"
#> 
#> 
#> 
#> $scores_raw
#> # A tibble: 250 × 3
#>    component location indicator
#>    <chr>     <chr>    <lgl>    
#>  1 cover     02       TRUE     
#>  2 diff      02       FALSE    
#>  3 repeats   02       FALSE    
#>  4 taper     02       FALSE    
#>  5 trend     02       FALSE    
#>  6 cover     04       FALSE    
#>  7 diff      04       FALSE    
#>  8 repeats   04       FALSE    
#>  9 taper     04       FALSE    
#> 10 trend     04       FALSE    
#> # ℹ 240 more rows
#>