Trend Component

This function identifies any change points in the forecast data or in the final observed data point. Change points are identified by any significant change in magnitude or direction of the slope of the time series.

Usage

plane_trend(location, input, seed, sig_lvl = 0.1)

Arguments

location

Character vector with location code; the location must appear in input and seed

input

Input signal data to be scored; object must be one of forecast or observed

seed

Prepared seed

sig_lvl

The significance level at which to identify change points (between zero and one); default is 0.1

Value

A list with the following values:

indicator: Logical as to whether or not the any forecast data or the final observed data point are a significant change point

output: An n x 7 tibble. The length of the forecast plus the observed data determine the length of n. The columns are:

Location: A character vector with the location code

Index: An integer index of all observed and forecast data

Date: The dates corresponding to all observed and forecast data (formatted as date)

Value: The incidence of all observed and forecast data (e.g., hospitalization rates)

Type: Indicates whether the data row is observed or forecast data

Changepoint: Logical identifying any change point (whether in observed or forecast data)

Flagged: Logical indicating whether or not the change point was flagged. Change points are only flagged if they are in the forecast data or are the final observed data point

flagged_dates: The date of any flagged change point(s). If there are none, NA is returned

Details

This function uses ecp::e.divisive(). Within e.divisive(), we use diff(x) instead of the raw data (x), which is a preference and slightly changes the way the points are identified. When we use diff(x), the index aligns with the gap between points rather than the points themselves. Instead of identifying a change point based on the change in size between two points, it identifies change points based on the change in the change itself. For example, the dataframe below shows an example of x and diff(x):

Index	x	diff(x)

1	3	6

2	9	0

3	9	28

4	37	37

5	74	1

6	75	0

7	75	0

Given this data, e.divisive(x) would identify index #5 (74) as the change point, because there was a jump of +37 between index 4 and 5. But e.divisive(diff(x)) would pick both index #3 (28) and #5 (1), because there was a jump of +28 from index 2 and 3, and there was a jump of -36 between index # 4 and 5. Ultimately, either way detects change points, but diff(x) seems to provide more information.

Further, we specify min.size = 2, which means that we are forcing a gap of at least 2 points between detecting change points. In a roundabout way, this increases the significance level or at least decreases the number of change points identified. Should we decide to change the function so that we're not using diff(x), it probably makes sense to change min.size to 3.

Examples

 ## read in example observed data and prep observed signal:
hosp <- read.csv(system.file("extdata/observed/hdgov_hosp_weekly.csv", package = "rplanes"))
tmp_hosp <-
  hosp %>%
  dplyr::select(date, location, flu.admits) %>%
  dplyr::mutate(date = as.Date(date))

prepped_observed <- to_signal(tmp_hosp, outcome = "flu.admits",
                             type = "observed", resolution = "weeks")

## read in example forecast and prep forecast signal:
prepped_forecast <- read_forecast(system.file("extdata/forecast/2022-10-31-SigSci-TSENS.csv",
                                               package = "rplanes")) %>%
   to_signal(., outcome = "flu.admits", type = "forecast", horizon = 4)

## prepare seed with cut date
prepped_seed <- plane_seed(prepped_observed, cut_date = "2022-10-29")

## Run plane trend component:
plane_trend(location = "05", input = prepped_forecast, seed = prepped_seed, sig_lvl = .2)
#> $indicator
#> [1] TRUE
#> 
#> $output
#> # A tibble: 20 × 7
#>    Location Index Date       Value Type     Changepoint Flagged
#>    <chr>    <int> <date>     <dbl> <chr>    <lgl>       <lgl>  
#>  1 05           1 2022-07-16    21 Observed FALSE       FALSE  
#>  2 05           2 2022-07-23    14 Observed FALSE       FALSE  
#>  3 05           3 2022-07-30     5 Observed TRUE        FALSE  
#>  4 05           4 2022-08-06     8 Observed FALSE       FALSE  
#>  5 05           5 2022-08-13     7 Observed FALSE       FALSE  
#>  6 05           6 2022-08-20     4 Observed FALSE       FALSE  
#>  7 05           7 2022-08-27     3 Observed FALSE       FALSE  
#>  8 05           8 2022-09-03     8 Observed FALSE       FALSE  
#>  9 05           9 2022-09-10     6 Observed FALSE       FALSE  
#> 10 05          10 2022-09-17     4 Observed FALSE       FALSE  
#> 11 05          11 2022-09-24     8 Observed FALSE       FALSE  
#> 12 05          12 2022-10-01    10 Observed TRUE        FALSE  
#> 13 05          13 2022-10-08    21 Observed FALSE       FALSE  
#> 14 05          14 2022-10-15    40 Observed FALSE       FALSE  
#> 15 05          15 2022-10-22    50 Observed FALSE       FALSE  
#> 16 05          16 2022-10-29    84 Observed TRUE        TRUE   
#> 17 05          17 2022-11-05    85 Forecast FALSE       FALSE  
#> 18 05          18 2022-11-12    90 Forecast FALSE       FALSE  
#> 19 05          19 2022-11-19    90 Forecast FALSE       FALSE  
#> 20 05          20 2022-11-26    92 Forecast FALSE       FALSE  
#> 
#> $flagged_dates
#> [1] "2022-10-29"
#> 
## Change location:
plane_trend(location = "09", input = prepped_forecast, seed = prepped_seed, sig_lvl = .2)
#> $indicator
#> [1] FALSE
#> 
#> $output
#> # A tibble: 20 × 7
#>    Location Index Date       Value Type     Changepoint Flagged
#>    <chr>    <int> <date>     <dbl> <chr>    <lgl>       <lgl>  
#>  1 09           1 2022-07-16     3 Observed FALSE       FALSE  
#>  2 09           2 2022-07-23     6 Observed FALSE       FALSE  
#>  3 09           3 2022-07-30     4 Observed FALSE       FALSE  
#>  4 09           4 2022-08-06     3 Observed FALSE       FALSE  
#>  5 09           5 2022-08-13     0 Observed FALSE       FALSE  
#>  6 09           6 2022-08-20     2 Observed FALSE       FALSE  
#>  7 09           7 2022-08-27     9 Observed FALSE       FALSE  
#>  8 09           8 2022-09-03     4 Observed FALSE       FALSE  
#>  9 09           9 2022-09-10     3 Observed FALSE       FALSE  
#> 10 09          10 2022-09-17     4 Observed FALSE       FALSE  
#> 11 09          11 2022-09-24     3 Observed FALSE       FALSE  
#> 12 09          12 2022-10-01     2 Observed FALSE       FALSE  
#> 13 09          13 2022-10-08     3 Observed FALSE       FALSE  
#> 14 09          14 2022-10-15     9 Observed FALSE       FALSE  
#> 15 09          15 2022-10-22     9 Observed TRUE        FALSE  
#> 16 09          16 2022-10-29    37 Observed FALSE       FALSE  
#> 17 09          17 2022-11-05    74 Forecast FALSE       FALSE  
#> 18 09          18 2022-11-12    75 Forecast FALSE       FALSE  
#> 19 09          19 2022-11-19    75 Forecast FALSE       FALSE  
#> 20 09          20 2022-11-26    75 Forecast FALSE       FALSE  
#> 
#> $flagged_dates
#> [1] NA
#> 
## Change sig_lvl:
plane_trend(location = "06", input = prepped_forecast, seed = prepped_seed, sig_lvl = .05)
#> $indicator
#> [1] FALSE
#> 
#> $output
#> # A tibble: 20 × 7
#>    Location Index Date       Value Type     Changepoint Flagged
#>    <chr>    <int> <date>     <dbl> <chr>    <lgl>       <lgl>  
#>  1 06           1 2022-07-16   102 Observed FALSE       FALSE  
#>  2 06           2 2022-07-23    64 Observed FALSE       FALSE  
#>  3 06           3 2022-07-30    46 Observed FALSE       FALSE  
#>  4 06           4 2022-08-06    39 Observed FALSE       FALSE  
#>  5 06           5 2022-08-13    38 Observed FALSE       FALSE  
#>  6 06           6 2022-08-20    39 Observed FALSE       FALSE  
#>  7 06           7 2022-08-27    33 Observed FALSE       FALSE  
#>  8 06           8 2022-09-03    35 Observed FALSE       FALSE  
#>  9 06           9 2022-09-10    29 Observed FALSE       FALSE  
#> 10 06          10 2022-09-17    24 Observed TRUE        FALSE  
#> 11 06          11 2022-09-24    35 Observed FALSE       FALSE  
#> 12 06          12 2022-10-01    72 Observed FALSE       FALSE  
#> 13 06          13 2022-10-08    93 Observed FALSE       FALSE  
#> 14 06          14 2022-10-15    97 Observed FALSE       FALSE  
#> 15 06          15 2022-10-22   124 Observed FALSE       FALSE  
#> 16 06          16 2022-10-29   211 Observed FALSE       FALSE  
#> 17 06          17 2022-11-05   249 Forecast FALSE       FALSE  
#> 18 06          18 2022-11-12   272 Forecast FALSE       FALSE  
#> 19 06          19 2022-11-19   288 Forecast FALSE       FALSE  
#> 20 06          20 2022-11-26   301 Forecast FALSE       FALSE  
#> 
#> $flagged_dates
#> [1] NA
#>