Read in forecast file

This function reads a probabilistic ("quantile") forecast csv file and prepares it for the to_signal function and downstream plausibility analysis. The quantile forecast file can be either a "legacy" or "hubverse" format (see Details for more information). The object returned is a tibble with summarized forecast data (i.e., prediction interval) for each location and horizon in the original file.

Usage

read_forecast(file, pi_width = 95, format = "legacy")

Arguments

file

Path to csv file containing quantile forecasts

pi_width

Width of prediction interval as integer; default 95 corresponds to 95% prediction interval

format

Format of the probabilistic format file; must be one of "legacy" or "hubverse" (see Details for more information); default is "legacy"

Value

A tibble with the following columns:

location: Geographic unit such as FIPS code

date: Date corresponding the forecast horizon

horizon: Forecast horizon

lower: Lower limit of the prediction interval for the forecast

point: Point estimate for the forecast

upper: Upper limit of the prediction interval for the forecast

Details

The probabilistic forecast format has been used by multiple forecasting hubs. In general, this format includes one row per combination of quantile, location, target, and horizon. At each row the forecasted value is provided. The specific format, including columns required, has changed over time. This function accommodates the "legacy" as well as more recent "hubverse" formats. For more details on specific columns and see the links in the References.

References

Hubverse: https://hubdocs.readthedocs.io/en/latest/user-guide/model-output.html

Legacy: https://github.com/cdcepi/Flusight-forecast-data/tree/master/data-forecasts#forecast-file-format

Examples

## read in example forecast and prep forecast signal (legacy format)
fp <- system.file("extdata/forecast/2022-10-31-SigSci-TSENS.csv", package = "rplanes")
read_forecast(fp)
#> # A tibble: 200 × 6
#>    location date       horizon lower point upper
#>    <chr>    <date>     <chr>   <dbl> <dbl> <dbl>
#>  1 02       2022-11-05 1           0     6    11
#>  2 02       2022-11-12 2           0     5    12
#>  3 02       2022-11-19 3           0     5    12
#>  4 02       2022-11-26 4           0     5    12
#>  5 04       2022-11-05 1          13    58   103
#>  6 04       2022-11-12 2           9    68   127
#>  7 04       2022-11-19 3           0    72   147
#>  8 04       2022-11-26 4           0    79   167
#>  9 05       2022-11-05 1          67    85   103
#> 10 05       2022-11-12 2          66    90   115
#> # ℹ 190 more rows

fp2 <- system.file("extdata/forecast/2023-11-04-SigSci-TSENS.csv", package = "rplanes")
read_forecast(fp2, format = "hubverse")
#> # A tibble: 204 × 6
#>    location date       horizon lower point upper
#>    <chr>    <date>       <dbl> <dbl> <dbl> <dbl>
#>  1 01       2023-11-04       0     0    33    69
#>  2 01       2023-11-11       1     0    34    84
#>  3 01       2023-11-18       2     0    34    94
#>  4 01       2023-11-25       3     0    35   103
#>  5 02       2023-11-04       0    18    29    40
#>  6 02       2023-11-11       1    13    28    43
#>  7 02       2023-11-18       2     9    27    45
#>  8 02       2023-11-25       3     6    26    46
#>  9 04       2023-11-04       0     0    56   124
#> 10 04       2023-11-11       1     0    63   173
#> # ℹ 194 more rows